Fixed exception in FileObject, a number indicating the run length followed by a $ end of row, means to insert n number of blank rows.
Fixed needing to return the number of digits in the StringMethods StringToInteger, as this is needed in order to know where to read the next line index character.

Added a Clear method to the engine base class, to be called when loading a new pattern.

Added additional code/features to the LifeForm for loading, pausing, clearing.